Yd 11

+++ app/Controllers/coursController.php
@@ -21,28 +21,26 @@ class CoursController extends Controller
         $this->cours = new Cours();
     }
 
-    // public function showcategoriesAndTags()
-    // {
-    //     $categories = $this->category->getCategories();
-    //     $tags = $this->tag->getTags();
-    //     // var_dump($tags);
-    //     $this->render('courses/index', [
-    //         'categories' => $categories,
-    //         'tags' => $tags
-    //     ]);
-    // }
 
     public function teacherCourses()
     {
-        $categories = $this->category->getCategories();
-        $tags = $this->tag->getTags();
-        // var_dump($tags);
-        $cours = $this->cours->getCourses($_SESSION['id']);
-        $this->render('courses/teacherCourses', [
-            'cours' => $cours,
-            'categories' => $categories,
-            'tags' => $tags
-        ]);
+        if ($_SESSION['role'] == 'teacher') {
+            $categories = $this->category->getCategories();
+            $tags = $this->tag->getTags();
+            $cours = $this->cours->getCourses($_SESSION['id']);
+            $pending = $this->cours->countPending($_SESSION['id']);
+            $accepted = $this->cours->countaccepted($_SESSION['id']);
+
+            $this->render('courses/teacherCourses', [
+                'cours' => $cours,
+                'categories' => $categories,
+                'tags' => $tags,
+                'pending' => $pending,
+                'accepted' => $accepted,
+            ]);
+        } else {
+            header('location: home');
+        }
     }
 
 
@@ -50,34 +48,38 @@ class CoursController extends Controller
 
     public function createCours()
     {
-        if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+        if ($_SESSION['role'] == 'teacher') {
+            if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 
-            $data = [
-                'title' => $_POST['title'],
-                'description' => $_POST['description'],
-                'content' => $_POST['content'],
-                'content_document' => $_POST['content_document'],
-                'content_video' => $_POST['content_video'],
-                'category' => $_POST['category'],
-                'level' => $_POST['level'],
-                'teacher' => $_SESSION['id']
-            ];
-            // var_dump($data);
-            // var_dump($_SESSION['id']);
+                $data = [
+                    'title' => $_POST['title'],
+                    'description' => $_POST['description'],
+                    'content' => $_POST['content'],
+                    'content_document' => $_POST['content_document'],
+                    'content_video' => $_POST['content_video'],
+                    'category' => $_POST['category'],
+                    'level' => $_POST['level'],
+                    'teacher' => $_SESSION['id']
+                ];
+                // var_dump($data);
+                // var_dump($_SESSION['id']);
 
 
-            if ($_POST['content'] === 'document') {
-                $this->cours->create($data);
-            } else if ($_POST['content'] === 'video') {
-                $this->cours->create($data, 'video');
+                if ($_POST['content'] === 'document') {
+                    $this->cours->create($data);
+                } else if ($_POST['content'] === 'video') {
+                    $this->cours->create($data, 'video');
+                }
+                $lastCourseId = $this->cours->getLastCourseId();
+                var_dump($lastCourseId);
+                if ($lastCourseId && isset($_POST['tags_id'])) {
+                    $this->cours->addTags($lastCourseId['id'], $_POST['tags_id']);
+                    header("Location: /teacherCourses");
+                    exit();
+                }
             }
-            $lastCourseId = $this->cours->getLastCourseId();
-            var_dump($lastCourseId);
-            if ($lastCourseId && isset($_POST['tags_id'])) {
-                $this->cours->addTags($lastCourseId['id'], $_POST['tags_id']);
-                header("Location: /teacherCourses");
-                exit();
-            }
+        } else {
+            header('location: home');
         }
     }
 
@@ -98,34 +100,43 @@ class CoursController extends Controller
 
     public function updateCourse()
     {
-        if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-            // var_dump($_POST['id']);
-            $data = [
-                'title' => $_POST['title'],
-                'description' => $_POST['description'],
-                'content' => $_POST['content'],
-                'content_video' => $_POST['content_video'],
-                'content_document' => $_POST['content_document'],
-                'category_id' => $_POST['category'],
-                'level' => $_POST['level']
-            ];
-            $this->cours->updateCourse($data, ['id' => $_POST['id']]);
-            $this->cours->deleteCourseTags($_POST['id']);
-            if (isset($_POST['tags_id'])) {
-                $this->cours->addTags($_POST['id'], $_POST['tags_id']);
-                header("Location: teacherCourses");
-                exit();
+        if ($_SESSION['role'] == 'teacher') {
+            if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+                // var_dump($_POST['id']);
+                $data = [
+                    'title' => $_POST['title'],
+                    'description' => $_POST['description'],
+                    'content' => $_POST['content'],
+                    'content_video' => $_POST['content_video'],
+                    'content_document' => $_POST['content_document'],
+                    'category_id' => $_POST['category'],
+                    'level' => $_POST['level'],
+                    'status' => 'pending'
+                ];
+                $this->cours->updateCourse($data, ['id' => $_POST['id']]);
+                $this->cours->deleteCourseTags($_POST['id']);
+                if (isset($_POST['tags_id'])) {
+                    $this->cours->addTags($_POST['id'], $_POST['tags_id']);
+                    header("Location: teacherCourses");
+                    exit();
+                }
             }
+        } else {
+            header('location: home');
         }
     }
 
     public function deleteCourse()
     {
-        if (isset($_GET['id'])) {
-            $id = ['id' => $_GET['id']];
-            $this->cours->deleteCourse($id);
-            header("location: teacherCourses");
-            exit();
+        if ($_SESSION['role'] == 'teacher') {
+            if (isset($_GET['id'])) {
+                $id = ['id' => $_GET['id']];
+                $this->cours->deleteCourse($id);
+                header("location: teacherCourses");
+                exit();
+            }
+        } else {
+            header('location: home');
         }
     }
 
@@ -154,16 +165,53 @@ class CoursController extends Controller
 
     public function showCoursesRequest()
     {
-        $result = $this->cours->showCoursesRequest();
-        $this->render(
-            'courses/coursesRequest',
-            ['request' => $result]
-        );
+        if ($_SESSION['role'] == 'teacher') {
+            $result = $this->cours->showCoursesRequest();
+            $this->render(
+                'courses/coursesRequest',
+                ['request' => $result]
+            );
+        } else {
+            header('location: home');
+        }
     }
 
     public function acceptCourse()
     {
-        $this->cours->acceptCourse($_GET['course'], $_GET['student']);
-        header('location: requests');
+        if ($_SESSION['role'] == 'teacher') {
+            $this->cours->acceptCourse($_GET['course'], $_GET['student']);
+            header('location: requests');
+        } else {
+            header('location: home');
+        }
+    }
+
+    public function getPendingCourses()
+    {
+        $request = $this->cours->getPendingCourses();
+
+        // print_r($request);
+        $this->render('courses/teacherCoursesRequest', [
+            'request' => $request
+        ]);
+    }
+
+    public function accept()
+    {
+        if ($_SESSION['role'] == 'admin') {
+            $this->cours->acceptTeacherCourse($_GET['id']);
+            header('location: coursesRequests');
+        } else {
+            header('location: home');
+        }
+    }
+    public function refuse()
+    {
+        if ($_SESSION['role'] == 'admin') {
+            $this->cours->rejectCourse($_GET['id']);
+            header('location: coursesRequests');
+        } else {
+            header('location: home');
+        }
     }
 }
